Clcok in init set up
