+++ app/DomainValueObjects/Log/Clock.php
@@ -1,22 +1,29 @@
 <?php 
 declare (strict_types = 1);
-namespace App\DomainValueObjects\Log\ClockIn;
+namespace App\DomainValueObjects\Log;
 
 use App\DomainValueObjects\UUIDGenerator\UUID;
 use App\DomainValueObjects\Log\TimeStamp\TimeStamp;
 use App\DomainValueObjects\Location\Location;
-
-use App\DomainValueObjects\Log\Clock;
 
 
 //Exceptions
 
 
 
-class ClockIn extends Clock
+class Clock
 {
+    private $timeStamp =  null;
+    private $location = null;
+
     public function __construct(UUID $uuid, TimeStamp $timeStamp, Location $location)
     {
-        parent::__construct($uuid, $timeStamp, $location);
+        $this->uuid = $uuid;
+        $this->timeStamp = $timeStamp;
+        $this->location = $location;
+        $this->validation();
     }
+
+    private function validation()
+    { }
 }
