+++ app/DomainValueObjects/Log/ClockOut/ClockOut.php
@@ -1,29 +1,23 @@
 <?php 
 declare (strict_types = 1);
-namespace App\DomainValueObjects\Log;
+namespace App\DomainValueObjects\Log\ClockOut;
 
 use App\DomainValueObjects\UUIDGenerator\UUID;
 use App\DomainValueObjects\Log\TimeStamp\TimeStamp;
 use App\DomainValueObjects\Location\Location;
+
+use App\DomainValueObjects\Log\Clock;
 
 
 //Exceptions
 
 
 
-class ClockIn
+class ClockOut extends Clock
 {
-    private $timeStamp =  null;
-    private $location = null;
 
     public function __construct(UUID $uuid, TimeStamp $timeStamp, Location $location)
     {
-        $this->uuid = $uuid;
-        $this->timeStamp = $timeStamp;
-        $this->location = $location;
-        $this->validation();
+        parent::__construct($uuid, $timeStamp, $location);
     }
-
-    private function validation()
-    { }
 }
